[Refactor] Split schema scanner modules and add UTs
