+++ be/src/exec/schema_scanner/schema_be_bvars_scanner.cpp
@@ -19,7 +19,6 @@
 
 #include "column/binary_column.h"
 #include "common/system/master_info.h"
-#include "exec/schema_scanner/schema_helper.h"
 #include "gutil/casts.h"
 #include "runtime/runtime_state.h"
 #include "types/logical_type.h"
