+++ be/src/exec/schema_scanner/schema_be_metrics_scanner.cpp
@@ -16,7 +16,7 @@
 
 #include "base/metrics.h"
 #include "common/system/master_info.h"
-#include "exec/schema_scanner/schema_helper.h"
+#include "exec/schema_scanner/schema_column_filler.h"
 #include "gutil/strings/substitute.h"
 #include "runtime/starrocks_metrics.h"
 #include "types/logical_type.h"
