+++ be/src/exec/schema_scanner/schema_be_threads_scanner.cpp
@@ -16,7 +16,7 @@
 
 #include "common/system/master_info.h"
 #include "common/thread/thread.h"
-#include "exec/schema_scanner/schema_helper.h"
+#include "exec/schema_scanner/schema_column_filler.h"
 #include "gutil/strings/substitute.h"
 #include "types/logical_type.h"
