+++ be/test/exec/CMakeLists.txt
@@ -43,3 +43,66 @@ target_link_libraries(starrocks_exec_core_test_objs PRIVATE StarRocksPCH)
 set_target_properties(starrocks_exec_core_test_objs PROPERTIES COMPILE_FLAGS "-fno-access-control")
 add_executable(exec_core_test $<TARGET_OBJECTS:starrocks_exec_core_test_objs>)
 target_link_libraries(exec_core_test ${EXEC_CORE_TEST_LINK_LIBS} gtest_main)
+
+set(SCHEMA_SCANNER_CORE_TEST_FILES
+    ../base/cxa_throw_wrap.cpp
+    schema_scanner/schema_scanner_core_test.cpp
+)
+
+set(SCHEMA_SCANNER_CORE_TEST_LINK_LIBS
+    ExecSchemaScannerCore
+    ExprCore
+    RuntimeCore
+    ChunkCore
+    ColumnCore
+    TypesCore
+    Common
+    Base
+    Gutil
+    StarRocksGen
+    ${STARROCKS_DEPENDENCIES}
+    ${STARROCKS_STATIC_LINK_LIBS}
+    ${WL_LINK_DYNAMIC} ${STARROCKS_SYSTEM_LINK_LIBS} ${WRAP_LINKER_FLAGS}
+    ${WL_START_GROUP}
+    gmock
+    gtest
+    ${WL_END_GROUP}
+)
+
+add_library(starrocks_schema_scanner_core_test_objs OBJECT ${SCHEMA_SCANNER_CORE_TEST_FILES})
+target_link_libraries(starrocks_schema_scanner_core_test_objs PRIVATE StarRocksPCH)
+set_target_properties(starrocks_schema_scanner_core_test_objs PROPERTIES COMPILE_FLAGS "-fno-access-control")
+add_executable(schema_scanner_core_test $<TARGET_OBJECTS:starrocks_schema_scanner_core_test_objs>)
+target_link_libraries(schema_scanner_core_test ${SCHEMA_SCANNER_CORE_TEST_LINK_LIBS} gtest_main)
+
+set(EXEC_SCHEMA_SCANNERS_TEST_FILES
+    ../base/cxa_throw_wrap.cpp
+    schema_scanner/exec_schema_scanners_test.cpp
+)
+
+set(EXEC_SCHEMA_SCANNERS_TEST_LINK_LIBS
+    ExecSchemaScanners
+    ExecSchemaScannerCore
+    RuntimeCore
+    ChunkCore
+    ColumnCore
+    TypesCore
+    Common
+    Base
+    Gutil
+    StarRocksGen
+    Util
+    ${STARROCKS_DEPENDENCIES}
+    ${STARROCKS_STATIC_LINK_LIBS}
+    ${WL_LINK_DYNAMIC} ${STARROCKS_SYSTEM_LINK_LIBS} ${WRAP_LINKER_FLAGS}
+    ${WL_START_GROUP}
+    gmock
+    gtest
+    ${WL_END_GROUP}
+)
+
+add_library(starrocks_exec_schema_scanners_test_objs OBJECT ${EXEC_SCHEMA_SCANNERS_TEST_FILES})
+target_link_libraries(starrocks_exec_schema_scanners_test_objs PRIVATE StarRocksPCH)
+set_target_properties(starrocks_exec_schema_scanners_test_objs PROPERTIES COMPILE_FLAGS "-fno-access-control")
+add_executable(exec_schema_scanners_test $<TARGET_OBJECTS:starrocks_exec_schema_scanners_test_objs>)
+target_link_libraries(exec_schema_scanners_test ${EXEC_SCHEMA_SCANNERS_TEST_LINK_LIBS} gtest_main)
